# docs/openapi/api-contract.locked.yaml
openapi: 3.0.3
info:
  title: trade-imports-address-book
  version: 1.0.0
  description: |
    Locked address-book API contract for EUDPA-58 (M0–M2).

    **Wire conventions:** camelCase JSON properties and query params; RFC 9457
    `application/problem+json` errors with camelCase `traceId` and (on validation
    failures) an `errors` map keyed by wire field name.

    **Identity:** every operation requires `Trade-Imports-Organisation-Id`. The path
    `{orgId}` must match the header value; mismatch returns **404** (no existence disclosure).

    **Soft delete:** DELETE sets internal status DELETED; the wire exposes `deleted: true`.
    Tombstones remain readable by id. A **404** is not a deletion signal.

    **Search:** `q` and `countryCode` combine with OR semantics when both are supplied.

    **Pagination:** page size is server-configured (default 25), not a request parameter.

    OperatorComplianceIT compares paths, HTTP methods, operationIds and component schema
    property names against this file.
servers:
  - url: http://localhost:8089
    description: Local dev
tags:
- name: addresses
  description: Address book, scoped to the caller's organisation
paths:
  /organisation/{orgId}/addresses/{operator-id}:
    get:
      tags:
      - addresses
      summary: Fetch one address, including tombstones
      operationId: get-address
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: Trade-Imports-Organisation-Id
        in: header
        required: true
        schema:
          type: string
      - name: operator-id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: The address (deleted false or true)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorResponse'
        '404':
          description: Unknown id, or an id outside the caller's organisation scope
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    put:
      tags:
      - addresses
      summary: Replace an address's mutable fields
      operationId: update-address
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: Trade-Imports-Organisation-Id
        in: header
        required: true
        schema:
          type: string
      - name: operator-id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
        required: true
      responses:
        '200':
          description: Updated address with bumped modifiedAt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorResponse'
        '400':
          description: Validation error, or a missing org header
          content:
            application/problem+json:
              schema:
                type: object
                anyOf:
                - $ref: '#/components/schemas/ValidationProblem'
                - $ref: '#/components/schemas/Problem'
        '404':
          description: Unknown id, an id outside the caller's organisation scope, or a tombstone
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      tags:
      - addresses
      summary: Soft-delete an address (tombstone)
      operationId: delete-address
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: Trade-Imports-Organisation-Id
        in: header
        required: true
        schema:
          type: string
      - name: operator-id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Soft-deleted (or already deleted — idempotent)
        '400':
          description: Missing org header
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Unknown id, or an id outside the caller's organisation scope
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /organisation/{orgId}/addresses:
    get:
      tags:
      - addresses
      summary: List the caller's ACTIVE addresses (paginated, searchable)
      operationId: list-operators
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: Trade-Imports-Organisation-Id
        in: header
        required: true
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: countryCode
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: One page of the caller's ACTIVE addresses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorPageResponse'
        '400':
          description: An out-of-range or non-numeric page, or a missing org header
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Path orgId does not match the caller's organisation header
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      tags:
      - addresses
      summary: Create an address in the caller's address book
      operationId: create-operator
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: Trade-Imports-Organisation-Id
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
        required: true
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: Relative URI of the created address
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorResponse'
        '400':
          description: Validation error, or a missing identity header
          content:
            application/problem+json:
              schema:
                type: object
                anyOf:
                - $ref: '#/components/schemas/ValidationProblem'
                - $ref: '#/components/schemas/Problem'
        '404':
          description: Path orgId does not match the caller's organisation header
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  schemas:
    AddressRequest:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 0
        addressLine1:
          type: string
          maxLength: 255
          minLength: 0
        addressLine2:
          type: string
          maxLength: 255
          minLength: 0
        townOrCity:
          type: string
          maxLength: 100
          minLength: 0
        county:
          type: string
          maxLength: 100
          minLength: 0
        postcode:
          type: string
          maxLength: 12
          minLength: 0
        countryCode:
          type: string
          maxLength: 2
          minLength: 1
        phone:
          type: string
          maxLength: 20
          minLength: 0
        email:
          type: string
          maxLength: 254
          minLength: 0
        type:
          type: object
        role:
          type: object
      required:
      - addressLine1
      - countryCode
      - email
      - name
      - phone
      - postcode
      - townOrCity
    ValidationProblem:
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        traceId:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    Problem:
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        traceId:
          type: string
    OperatorResponse:
      properties:
        id:
          type: string
        name:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        townOrCity:
          type: string
        county:
          type: string
        postcode:
          type: string
        countryCode:
          type: string
        phone:
          type: string
        email:
          type: string
        organisationId:
          type: string
        deleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
    OperatorPageResponse:
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OperatorResponse'
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalItems:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32